requirements.txt
pytest
PyGithub
statistics
matplotlib
numpy